server/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
httpx>=0.26.0

# Data & Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
alembic>=1.13.0

# Vector Search (optional for prototype)
numpy>=1.26.0
sentence-transformers>=2.2.0
torch>=2.0.0

# File Processing
python-magic>=0.4.27
pypdf>=3.17.0
python-docx>=1.1.0